implemented new tokenizer
replaces Template#tokenize method with a full tokenizer class. this class takes
the raw liquid source and returns each token using the #next_token or #shift
methods. tokenizer is aware of the line number of the most recently fetched
token which can be useful for giving users more meaningful feedback when the
make a mistake like a syntax error.

modified Liquid::SyntaxError to optionally include a token so a developer can
include the line number which the error occurred in error messages.